requirements.txt
attrs
scipy
numpy
pyyaml
